Docs, keithley2400, and __init__ fixes - Updated docs to reflect current set of drivers - Fixed __init__ files to properly import classes - Changed Keithley 2400 class to make it easier to configure and to prevent it from being locked if changes needed to be made on the device itself.
