+++ docs/labdrivers.keithley.rst
@@ -2,10 +2,10 @@
 labdrivers.keithley 
 ##########################
 
+.. toctree::
+    :maxdepth: 1
+
+
 .. autoclass:: labdrivers.keithley.keithley2400.keithley2400
     :members:
     :undoc-members:
-
-
-.. toctree::
-    :maxdepth: 1
