+++ labdrivers/lakeshore/__init__.py
@@ -1,0 +1,1 @@
+from .ls332 import ls332
