+++ labdrivers/oxford/__init__.py
@@ -1,2 +1,4 @@
 from .ips120 import ips120
 from .itc503 import itc503
+from .mercuryips import mercuryips
+from .triton200 import triton200
